refactor: code review changes to simplify types
