+++ v3/src/components/graph/components/casedots.tsx
@@ -1,7 +1,8 @@
 import {randomUniform, select} from "d3"
 import {onAction} from "mobx-state-tree"
 import React, {useCallback, useEffect, useRef, useState} from "react"
-import {CaseData, IDotsRef} from "../graphing-types"
+import {CaseData} from "../d3-types"
+import {IDotsRef} from "../graphing-types"
 import {ICase} from "../../../models/data/data-set-types"
 import {isAddCasesAction} from "../../../models/data/data-set-actions"
 import {useDragHandlers, usePlotResponders} from "../hooks/use-plot"
