+++ v3/src/components/graph/components/chartdots.tsx
@@ -1,7 +1,7 @@
 import {ScaleBand} from "d3"
 import React, {useCallback} from "react"
-import { selectDots } from "../d3-types"
-import {attrRoleToAxisPlace, CaseData, PlotProps} from "../graphing-types"
+import {CaseData, selectDots} from "../d3-types"
+import {attrRoleToAxisPlace, PlotProps} from "../graphing-types"
 import {usePlotResponders} from "../hooks/use-plot"
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context"
 import {useDataSetContext} from "../../../hooks/use-data-set-context"
