+++ v3/src/components/graph/components/dotplotdots.tsx
@@ -1,7 +1,8 @@
 import {max, range, ScaleBand, ScaleLinear, select} from "d3"
 import {observer} from "mobx-react-lite"
 import React, {useCallback, useRef, useState} from "react"
-import {CaseData, PlotProps} from "../graphing-types"
+import {CaseData} from "../d3-types"
+import {PlotProps} from "../graphing-types"
 import {useDragHandlers, usePlotResponders} from "../hooks/use-plot"
 import {appState} from "../../../models/app-state"
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context"
