+++ v3/src/components/graph/components/scatterdots.tsx
@@ -2,7 +2,8 @@ import {ScaleBand, ScaleLinear, select} from "d3"
 import React, {useCallback, useRef, useState} from "react"
 import {appState} from "../../../models/app-state"
 import {ScaleNumericBaseType} from "../../axis/axis-types"
-import {CaseData, PlotProps} from "../graphing-types"
+import {CaseData} from "../d3-types"
+import {PlotProps} from "../graphing-types"
 import {useDragHandlers, usePlotResponders} from "../hooks/use-plot"
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context"
 import {useDataSetContext} from "../../../hooks/use-data-set-context"
