+++ v3/src/components/graph/graphing-types.ts
@@ -1,9 +1,7 @@
 import React from "react"
 import {AxisPlace} from "../axis/axis-types"
 import {GraphPlace} from "../axis-graph-shared"
-
- // The data stored with each plot element (e.g. 'circle')
-export type CaseData = { plotNum: number, caseID: string }
+import {DotsElt} from "./d3-types"
 
 export const PrimaryAttrRoles = ['x', 'y'] as const
 export type PrimaryAttrRole = typeof PrimaryAttrRoles[number]
@@ -14,7 +12,7 @@ export const GraphAttrRoles = [
 export type GraphAttrRole = typeof GraphAttrRoles[number]
 export type IsGraphDropAllowed = (place: GraphPlace, attrId?: string) => boolean
 
-export type IDotsRef = React.RefObject<SVGSVGElement | null | undefined>
+export type IDotsRef = React.MutableRefObject<DotsElt>
 
 export const attrRoleToAxisPlace: Partial<Record<GraphAttrRole, AxisPlace>> = {
   x: "bottom",
